backend/pyproject.toml
[project]
name = "threat-intel"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiohttp>=3.13.3",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.3",
    "google-genai>=1.59.0",
    "ipython>=9.9.0",
    "loguru>=0.7.3",
    "mlflow>=3.8.1",
    "nest-asyncio>=1.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "playwright>=1.57.0",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "sqlalchemy[asyncio]>=2.0.45",
    "supabase>=2.28.0",
    "tenacity>=9.1.2",
]
